pyproject.toml
[project]
name = "simulador-entrevistas"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.9",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "motor (>=3.7.1,<4.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (<4.1)",
    "numpy (==2.2.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "simulador_entrevistas", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
